+++ packages/composites/rocket-pool/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @chainlink/rocket-pool-adapter
 
+## 1.1.4
+
+### Patch Changes
+
+- Updated dependencies [2fdaa5aa4]
+  - @chainlink/ea-bootstrap@2.26.1
+  - @chainlink/ea-reference-data-reader@1.3.28
+
 ## 1.1.3
 
 ### Patch Changes