+++ packages/sources/coinmetrics-test/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @chainlink/coinmetrics-test-adapter
 
+## 1.2.1
+
+### Patch Changes
+
+- 2fdaa5aa4: Bump v3 version
+
+## 1.2.0
+
+### Minor Changes
+
+- ef068e559: Added burned endpoints and price http endpoint
+
+### Patch Changes
+
+- bc8886c63: Bumped framework version
+
 ## 1.1.1
 
 ### Patch Changes