+++ packages/sources/dns-query/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @chainlink/dns-query-adapter
 
+## 1.6.21
+
+### Patch Changes
+
+- Updated dependencies [2fdaa5aa4]
+  - @chainlink/ea-bootstrap@2.26.1
+  - @chainlink/ea-test-helpers@1.4.2
+
 ## 1.6.20
 
 ### Patch Changes